requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
async-timeout==5.0.1
backports-datetime-fromisoformat==2.0.3
blinker==1.9.0
certifi==2026.1.4
cffi==1.17.1
click==8.2.1
cryptography==45.0.4
Flask==3.1.1
Flask-JWT-Extended==4.7.1
Flask-SocketIO==5.3.6
flask-marshmallow==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
gunicorn==25.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==4.2.1
marshmallow-sqlalchemy==1.4.2
minio==7.2.20
packaging==26.0
pycparser==2.22
pycryptodome==3.23.0
PyJWT==2.10.1
python-dotenv==1.2.1
redis==6.2.0
SQLAlchemy==2.0.41
psycopg2-binary==2.9.9
typing_extensions==4.14.0
urllib3==2.6.3
Werkzeug==3.1.3

Flask-Cors==4.0.1
eventlet==0.40.3
